Bulk move everything into a Testwork Extension
