+++ src/Behat/Behat/Tester/ServiceContainer/TesterExtension.php
@@ -48,8 +48,6 @@ class TesterExtension extends BaseExtension
     const BACKGROUND_TESTER_WRAPPER_TAG = 'tester.background.wrapper';
     const STEP_TESTER_WRAPPER_TAG = 'tester.step.wrapper';
 
-    const PRIORITISER_TAG = 'tester.prioritiser';
-
     /**
      * @var ServiceProcessor
      */
@@ -96,9 +94,6 @@ class TesterExtension extends BaseExtension
         parent::load($container, $config);
 
         $this->loadRerunController($container, $config['rerun_cache']);
-        $this->loadPriorityController($container);
-        $this->loadPrioritisingExercise($container);
-        $this->loadDefaultPrioritisers($container);
         $this->loadPendingExceptionStringer($container);
     }
 
@@ -114,7 +109,6 @@ class TesterExtension extends BaseExtension
         $this->processExampleTesterWrappers($container);
         $this->processBackgroundTesterWrappers($container);
         $this->processStepTesterWrappers($container);
-        $this->processPrioritisers($container);
     }
 
     /**
@@ -237,67 +231,6 @@ class TesterExtension extends BaseExtension
         ));
         $definition->addTag(CliExtension::CONTROLLER_TAG, array('priority' => 200));
         $container->setDefinition(CliExtension::CONTROLLER_TAG . '.rerun', $definition);
-    }
-
-    /**
-     * Loads priority controller.
-     *
-     * @param ContainerBuilder $container
-     */
-    protected function loadPriorityController(ContainerBuilder $container)
-    {
-        $definition = new Definition('Behat\Behat\Tester\Cli\PriorityController', array(
-            new Reference(EventDispatcherExtension::DISPATCHER_ID),
-            new Reference(self::EXERCISE_WRAPPER_TAG . '.prioritising')
-        ));
-        $definition->addTag(CliExtension::CONTROLLER_TAG, array('priority' => 250));
-        $container->setDefinition(CliExtension::CONTROLLER_TAG . '.priority', $definition);
-    }
-
-    /**
-     * Loads exercise wrapper that enables prioritisation
-     *
-     * @param ContainerBuilder $container
-     */
-    protected function loadPrioritisingExercise(ContainerBuilder $container)
-    {
-        $definition = new Definition('Behat\Behat\Tester\Priority\Exercise', array(
-            new Reference(self::EXERCISE_ID)
-        ));
-        $definition->addTag(self::EXERCISE_WRAPPER_TAG, array('priority' => -9999));
-        $container->setDefinition(self::EXERCISE_WRAPPER_TAG . '.prioritising', $definition);
-    }
-
-    /**
-     * Defines default prioritisers
-     *
-     * @param ContainerBuilder $container
-     */
-    protected function loadDefaultPrioritisers(ContainerBuilder $container)
-    {
-        $definition = new Definition('Behat\Behat\Tester\Priority\Prioritiser\ReversePrioritiser');
-        $definition->addTag(self::PRIORITISER_TAG, array('priority' => -9999));
-        $container->setDefinition(self::EXERCISE_WRAPPER_TAG . '.reverse', $definition);
-
-
-        $definition = new Definition('Behat\Behat\Tester\Priority\Prioritiser\RandomPrioritiser');
-        $definition->addTag(self::PRIORITISER_TAG, array('priority' => -9999));
-        $container->setDefinition(self::EXERCISE_WRAPPER_TAG . '.random', $definition);
-    }
-
-    /**
-     * Dynamically registers tagged prioritisers
-     *
-     * @param ContainerBuilder $container
-     */
-    protected function processPrioritisers(ContainerBuilder $container)
-    {
-        $definition = $container->getDefinition(CliExtension::CONTROLLER_TAG . '.priority');
-        $references = $this->processor->findAndSortTaggedServices($container, self::PRIORITISER_TAG);
-
-        foreach ($references as $reference) {
-            $definition->addMethodCall('registerPrioritiser', array($reference));
-        }
     }
 
     /**
